Move root options into separate package
