+++ options/root/flags.go
@@ -14,7 +14,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-package options
+package root
 
 import (
 	"flag"
