+++ options/root/root_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package options
+package root
 
 // TODO(tests): Uncomment tests once Codecov is working.
 /*
